// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  auth: typeof auth;
  emails: typeof emails;
  http: typeof http;
  projects: typeof projects;
  stripe: typeof stripe;
  stripeConfig: typeof stripeConfig;
  userHelpers: typeof userHelpers;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;